requirements.txt
streamlit
google-genai
psycopg2-binary
pandas
sqlalchemy
pg8000
python-dotenv
google-cloud-storage
requests